Abdou

+++ FitnessApp/src/screens/CreateRoutine.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { View, Text, TextInput, Button, StyleSheet, TouchableOpacity } from 'react-native';
+import { View, Text, TextInput, StyleSheet, TouchableOpacity, Alert } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 import { Routine, Exercise, Set } from './ViewRoutine';
@@ -59,21 +59,31 @@ const CreateRoutine = ({ navigation }: Props) => {
     try {
       const existingRoutines = await AsyncStorage.getItem('routines');
       const parsedRoutines = existingRoutines ? JSON.parse(existingRoutines) : [];
+  
+      // Check if routine name already exists
+      const routineExists = parsedRoutines.some((routine: Routine) => routine.name === routineName);
+  
+      if (routineExists) {
+        Alert.alert('Routine name already exists');
+        return;
+      }
+      // Save routine 
       const updatedRoutines = [...parsedRoutines, newRoutine];
       await AsyncStorage.setItem('routines', JSON.stringify(updatedRoutines));
-      console.log('Routine saved successfully:', newRoutine);
-  
-      setTimeout(() => {
-        navigation.navigate('ViewRoutine');
-      }, 500);
+      console.log('Routine saved successfully:321', newRoutine);
+      // Go to viewRoutine screen
+      navigation.navigate('ViewRoutine');
     } catch (e) {
       console.error('Error saving routine:', e);
     }
   };
-  
-  
-  
 
+  const handleCancel = () => {
+    navigation.navigate('ViewRoutine');
+  };
+  
+  
+  
   return (
     <View style={styles.container}>
       <View style={styles.inputContainer}>
@@ -102,6 +112,9 @@ const CreateRoutine = ({ navigation }: Props) => {
         </TouchableOpacity>
         <TouchableOpacity style={styles.saveButton} onPress={handleSaveRoutine}>
           <Text style={styles.buttonText}>Save Routine</Text>
+        </TouchableOpacity>
+        <TouchableOpacity style={styles.cancelButton} onPress={handleCancel}>
+          <Text style={styles.buttonText}>Cancel</Text>
         </TouchableOpacity>
       </View>
 
@@ -153,6 +166,16 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     elevation: 5,
     paddingVertical: 8,
+    marginBottom: 8,
+  },
+  cancelButton: {
+    borderRadius: 10,
+    backgroundColor: 'red',
+    alignItems: 'center',
+    justifyContent: 'center',
+    elevation: 5,
+    paddingVertical: 8,
+    marginBottom: 8,
   },
   buttonText: {
     color: '#fff',
